refactor: more moves & rename RootObject -> RootEntry
