+++ crates/ltk_ritobin/src/ast/node/kind.rs
@@ -1,0 +1,7 @@
+#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq)]
+pub enum NodeKind {
+    Object,
+    Struct,
+    Property,
+    Value,
+}
